refactor(pipeline): extract review.go from engine.go [#288]
Move parallel-review orchestration code (runParallelReview, runReviewer,
runReviewerWithRetry, mergeReviewFindings, computeReviewVerdict,
buildReviewArtifact, loadPriorReview, neededReviewersFromPrior,
priorFindingsForReviewer) and supporting types (reviewerResult,
reviewerMsg, reviewerLog) into a dedicated review.go file.
